+++ src/components/Typography/index.ts
@@ -1,3 +1,3 @@
-// Typography from copy-ui @ 2025-10-02.
+// Typography from copy-ui @ 2025-10-03.
 
 export { Typography } from "./Typography";