+++ src/components/Typography/CHANGELOG.md
@@ -1,1 +1,4 @@
-- 2025-10-02: Initial version (with H1 component only).
+- 2025-10-02: Initial version (with Typography.H1 component).
+- 2025-10-03: Added Typography.H2, Typography.H3 and Typography.P components.
+  Support the `mt` prop for margin-top spacing & update CSS styles to make it
+  look better.